detect-tls-ja3-hash: move unittests to tests/
